\name{adjClustFast-package}
\alias{adjClustFast-package}
\alias{adjClustFast}
\docType{package}
\title{
\packageTitle{adjClustFast}
}
\description{
\packageDescription{adjClustFast}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adjClustFast}
\packageIndices{adjClustFast}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{adjClustFast}

Maintainer: \packageMaintainer{adjClustFast}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
